Add package and nuget search category.
Variation to the existing add package command in the unified
search that uses the tags "package" or "nuget" and allows a
package to be installed. Examples below:

    nuget:automapper
    nuget:automapper 1.2.3
    package:Xamarin.Forms
    package:Xamarin.Forms 1.2.3

With the above typed in the unified search the user can select
Add Package to install the package.

The global Add Package command is hidden when the package or nuget
tag are being used.