Prevent uninstall.ps1 scripts stopping packages being uninstalled.
The uninstall.ps1 script is run before the package is removed. If an
exception occurs the uninstall will fail. Any exceptions thrown
during install or uninstall are not caught and logged, but do not
cause the main install or uninstall to fail. These exceptions are
mainly because Pash does not fully implement every PowerShell feature.