Prevent exceptions in search categories crashing IDE.
If the ItemCount property of the ISearchDataSource throws an exception
then when anything is typed into the global search text box (top right
of IDE) the IDE crashes when not on Windows. This can happen if the
version of NuGet.Core used by the IDE is different to the one used by
the addin.