Allow console to be cleared using command line.
The PowerShell host now redefines the Clear-Host command so it clears
the console window instead of showing an "Handle is invalid" error.
Clear-Host or Clear can now be used in the PowerShell console as a
command instead of having to click the Clear button.

Pash does not currently support the $host variable so instead the
Get-Host command is used.