Allow Package Console to work when no solution is open
Fixed loader exceptions when the Package Console is opened before any
solution is opened by ensuring the NuGet.Packaging.dll and the
NuGet.Packaging.Core.Types.dll are copied to the addin directory.

Fix null reference when trying to get projects for Get-Package when no
solution is open.

Ensure package source list is populated when no solution is open.