Run PowerShell scripts when adding, removing package
PowerShell scripts are now run when adding or removing a NuGet package
using the PowerShell cmdlets. Currently the PowerShell scripts are
only run for projects that use a packages.config file.

init.ps1 is not currently run on opening the solution. It will also be
run on every install of a NuGet package into the same solution when it
should only run once.