Add browse, installed, update and consolidate links to manage packages dialog
Currently these links do not switch the package results but only show the
current page is selected by underlining the link and marking the text bold.
The links are labels and are added between the package source combo box and
the search text box. It is not currently possible to put them above the
package source combo box since there is a limit on how tall the toolbar can
be and you end up with the toolbar background image being not tall enough and
does not cover the background of all the widgets at the top of the dialog.