Can now use Manage Packages dialog for a single project
Added a Manage Packages menu to the context menu shown when you
right click the Packages folder. Selecting this menu will open the
Manage Packages dialog and will allow you to manage the packages for
that single project. The Consolidate tab is not shown. The project
selection dialog is also not shown.