Fix null reference when adding NuGet packages using PowerShell console
The solution manager has a new SaveProject method and the NuGet addin
does not expect the NuGet project's MSBuildNuGetProjectSystem to
be a different type so a null reference is thrown when this method
is called. To handle this the NuGet addin's MonoDevelopSolutionManager
is wrapped with a NuGet extension solution manager class that
delegates everything to the wrapped solution manager apart from the
SaveProject method.