Fix null reference when running init.ps1
The tools path is added to the current path environment variable but
a lowercase 'path' was being searched for which returned a null
value causing a null reference exception. Now the code looks for a
'PATH' environment variable which should be found and also handles
a null value being returned.