Addin now works with preview 7 (7.0 Build 2740)
1. Updated the MonoDevelop and NuGet binaries.
2. Updated implementations of IMonoDevelopSolutionManager and
INuGetPackageManager due to new methods and events.
3. Strong named the extensions and cmdlet projects. The extensions
project needs to be strong named in order to access the main
NuGet addin since it is strong named. The cmdlet project needs to
be strong named in order to access the extensions internals. The
PowerShell and Pash project seem to work without being strong
named.
4. Updated the addin and assembly version to 0.12.2