Only show license acceptance dialog once for multiple projects
The license acceptance dialog was shown for each project when a NuGet
package was installed, updated or consolidated. Now the check is only
done at the start when installing into the first project. If the
licenses are declined then no packages will be installed since all
actions are queued together.