Add script to launch the NuGet cmdlet with PowerShell core
Then Visual Studio for Mac can attach to pwsh to debug the cmdlets.
Currently not possible to run pwsh via a project run config and
debug the cmdlets. When running via a project run config the IDE does
not run the debugger and also does not launch pwsh in an external
console so it is not possible to enter commands.