Send active package source to remote PowerShell host
Currently Find-Package fails if the active package source is not in the
~/.nuget/NuGet/NuGet.Config file. This is used by the PowerShell
remote host since it is running with .NET Core and not Mono. So the
package sources defined in the IDE do not currently match those read
by NuGet. The NuGet extensions addin will have to send its full
list of package sources to the remote PowerShell host.