Add assembly resolution failure logging to PowerShell console host
Lucene.NET tries to use System.Configuration.ConfigurationManager and
this can be seen by checking the exception of the $error in PowerShell
by using:

  $error[0].Exception.ToString()

Use the AssemblyLoadContext.Default.Resolving event to see the
assemblies that cannot be found for the Default AssemblyLoadContext.
This information is logged by the remote PowerShell console host.