Fix PowerShell prompt being displayed before output messages
The first time a PowerShell command was run the 'PM> ' prompt would
be displayed followed by output from the command. This would then
result in the output being used as the next command when the return
key was pressed. To avoid this the output from the remote PowerShell
console host is no longer using the notify method with Json RPC which
does not wait for a response. Instead the invoke method is used. This
ensures the message has been received and should ensure all output is
handled before the PowerShell command finishes running.