Fix PowerShell console width not known in remote host
The first time the PowerShell console host is opened the max column
width is now sent to the remote host. Previously the remote host
was stuck at 80 columns until the IDE window was resized so the
output would be wrapped even when there was enough width to display
the output completely unwrapped.