Implement Get-Package
Had a problem with the original async await code with the
WriteUpdatePackagesFromRemoteSourceAsync method calling WriteObject
on a different thread from the thread used in ProcessRecord. This
results in PowerShell throwing a PSInvalidOperationException. To
avoid this the WriteUpdatePackagesFromRemoteSourceAsync method
now returns the package information and the result of this is
used by blocking on the Task's Result.