Fix error about missing Invoke-InitializePackages cmdlet
Do not call this cmdlet. Currently this cmdlet is not available
for the remote powershell host and may never be depending on the
implementation.

Also fixed the working directory being set on opening and closing
a solution for the PowerShell host. Set-Location is called with
the solution's directory when a solution is opened. When the
solution is closed the Set-Location is reset back to the HOME
or USERPROFILE directory, depending on the operating system being
used.