Fix null reference on opening PowerShell console host the first item
The width of the window can change before the first PowerShell command
has been run. This means the JsonRpc instance has not yet been created
in the remote PowerShell host so a null reference would occur. This
null JsonRpc instance is now checked for. After the first PowerShell
command is run on initializing the console host the number of visible
columns is updated so ignoring the event on opening the window the first
time does not mean the remote PowerShell host has the wrong column
width.