Fix aggregate exceptions logged and also displayed in PowerShell host
If an error occurred whilst processing a cmdlet json rpc request in
the main NuGet extensions addin an AggregateException was being
generated. This aggregate exception was being logged and also resulted
in the 'One or more errors occurred' message being included in the
error shown in the PowerShell console. To avoid this the code in the
main NuGet extensions addin is using GetAwaiter ().GetResult () instead
of Wait () or .Result, which prevents the AggregateException from
being thrown. Also fixed some incorrect logging messages which had
string formatting placeholders but were not being used.