Add Sync-Package PowerShell command
Unfortunately the implementation of this command seems to be wrong,
even in Visual Studio on Windows. It seems that sync-package is
supposed to be the command line equivalent of the consolidate feature
available in the UI. However sync-package will install the specified
NuGet package into all other projects, even if it was not installed
before, which does not seem correct.