PowerShell Console window now has a stop button
This will set a token to be cancelled. This token is used in the main
addin when using NuGet APIs that support a cancellation token.
Currently this does not stop commands running in the PowerShell
remote host - only affects those calls from the host made to the
main addin via Json RPC.