Support stopping the PowerShell pipeline in the remote host
When the Stop button is pressed now a message is sent to the remote
PowerShell console host to stop the running command. This will set
a cancellation token to be cancelled, in case any NuGet requests are
being run directly via PowerShell, and also the pipeline will be stopped.
The pipeline completion reason is checked and if there is an exception
this is reported in the console window. When stopping the pipeline this
results in an 'The pipeline has been stopped' error displayed in the
console window the same as what happens with Visual Studio on Windows.