Add button to Package Console window to configure sources
Added a button next to the list of package sources which when
clicked will open the NuGet package sources in the preferences
dialog. This behaviour is different to the Add Packages dialog
which has a Configure Sources... item in the list of sources
which will open the preferences dialog. Visual Studio on Windows
uses a separate button which matches what is implemented here.