Support running init, install and uninstall PowerShell scripts
The init.ps1, install.ps1 and uninstall.ps1 scripts are now
run when installing and uninstalling packages in the package console.
The init.ps1 is not currently run for PackageReference projects and
is also not run on opening the solution when an installed package
exists that uses an init.ps1 file.

Also seeing an extra newline after each Write-Host call in any
powershell scripts included in a NuGet package.