Fix extra new lines when using Write-Host in PowerShell scripts
Handle the Write method being used on the PowerShell UI host.
Here if the message does not contain a newline the message is
added to a string builder until a newline is received at the end
of the text. Only then is the message sent to the console window
in the IDE.