Fix output when running init.ps1 being on the PowerShell prompt line
On opening the PowerShell console with a project that uses an init.ps1
file the output from running this init.ps1 file would begin on the
PowerShell PM> prompt line. Now a new line is written to the console
window to avoid this.

Also now the init.ps1 script is run with the NuGet background dispatcher.
The stop button is enabled when the init.ps1 is being run on solution
open.