Fix duplicate PowerShell prompts
On opening the PowerShell console with a solution open the
PowerShell PM> prompt would be displayed twice. The problem was
that a check is made to see if any init.ps1 scripts need to be run
and this then caused a second prompt to be displayed even when
no scripts were run.