Fix Copy keyboard shortcut not working in PowerShell console
For some reason the default copy command handler does not work. Not sure
why since it seems OK in the ImmediatePad which is similar to the
PowerShell console. As a workaround the PackageConsoleView now has a
command handler for EditCommands.Copy and uses similar code to the IDE's
DefaultCopyHandler to copy text.