Re-use main ConsoleView
Removed the ConsoleView2 and try to use the IDE's main ConsoleView.
This is mostly possible. ConsoleView2 was adding new text colour tags
for errors, warnings and debug information. Only problem is the
scriptLines string which is not available to clear if you want to
clear the console without automatically re-generating the prompt.
Need to see if the ConsoleView's Clear method can be used instead.