Implement EnvDTE ActiveConfiguration Properties for project
This allows the OutputPath to be retrieved for the active configuration.
This is done with a Json RPC call back to the main NuGet extensions
addin which then finds the active configuration and tries to find
the property being requested. Also will convert backslashes to
forward slashes for the OutputPath since by default this property
value will be the MSBuild property in the project which uses
backslashes.