Implement parts of IVsBrowseObjectContext for .NET Core projects
If the project's target framework is .NETCoreApp or .NETStandard then
when using EnvDTE a CpsProject type is returned which implements the
IVsBrowseObjectContext interface. This allows EF Core tools to get
a little further. EF Core's NuGet module uses the IVsBrowseObjectContext
to get the ProjectAssetsFile path which is available as an MSBuild
property.