Allow EF Core to find the language for .NET Core projects
EF Core tools was getting the Language by using MSBuild's
GlobalProjectCollection. Currently this is not available. To avoid
this being called for .NET Core projects the PackageUtilities has
been changed so that it determines the project's are CPS projects.
This avoids MSBuild being used. For .NET Framework projects MSBuild
will be used so an implementation of GlobalProjectCollection will
be needed for non .NET Core projects.

This allows EF core to get to the point where it attempts to run
dotnet to generate a migration when using Add-Migration.