Add empty EnvDTE ProjectItems implementation
This is enough to allow EF Core to get a bit further adding a
migration for a .NET Core project. It is using the ProjectItems
AddFromFile method but since .NET Core projects use wildcards
the new file does not need to be explicitly added to the project
so an empty implementation is OK here for now.