Support EF Script-Migration
Convert a request for a .NET Core project's IntermediateOutputPath
to the same request to the active project configuration. Using the
IntermediateOutputPath property from the project returns an obj
directory with guids as part of the path. This directory does not
exist so the PowerShell Join-Path failed to resolve it. Using the
active configuration returns the obj/Debug/netcoreapp2.1 directory.