Call PowerShell's TabExpansion2 to find items
Instead of having a hard coded list of NuGet package commands when
the tab key is pressed in the PowerShell console a request is sent
via JsonRpc to the remote host and it uses TabExpansion2 to find
any expansions. Currently only the default expansions provided by
PowerShell are available. There are no NuGet specific tab expansions
currently available.