Allow unique project name to be used in PowerShell console
EF Core uses the unique project name when calling Get-Project for
some of its tab expansions. To support this when looking for a project
given a name a check is made for a match with the project name and
its unique name. This allows EF Core's tab expansion to work when
it looks for a list of providers.