PowerShell cmdlet can now prompt for choices
This allows EF Core's Drop-Database to prompt the user to ask if the
database should be dropped. The PowerShell window will show the message
and the user can enter an option and press return.

Currently cancelling is not well supported. The stop button in the
pad window will stop the pipeline but the prompt will remain until some
text is entered.