Fix problem using EnvDTE.ProjectItems with Analyzer items
Analyzer project items return true for IsLink but the Link property
itself is not set. They are treated as links since the include path
is outside the project. This was causing an error when trying to
get the directory name using Path.GetDirectoryName since the link
has no value. To avoid this the Link property is checked and if this
is null or empty the ProjectFile's FilePath is used instead.