Restart RemotePowerShell host if it crashes
If the RemotePowerShell host crashes it is now restarted and the JsonRpc
channel is recreated. Note that on a crash the Process.Exited event is
not fired. The JsonRpc.Disconnected event is fired with the event
args showing a reason of Disposed. This Disconnected event is used to
trigger the remote PowerShell host being started again and the JsonRpc
connection re-initialized. An error is displayed in the PowerShell
console window when the disconnect occurs.