Prevent repeated attempts to restart PowerShell host console app
Add a simple backoff strategy when restarting the PowerShell host
console app fails. If the Json rpc connection is disconnected then
check the last time a restart occurred. If this occurred within 5
seconds then do not attempt to restart the remote PowerShell host
console. This prevents getting into a loop where the NuGet Package
Manager Console is repeatedly trying to restart the process if
it is failing every time.