Improve PowerShell console tab completion
The tab completion window would be closed on typing in another
character after it was opened. Now typing more characters will
filter the list of items.

Note that the completion window behaviour is not perfect. Certain
characters, such as '-', will cause the window to close even if
that would match an item in the list.