Fix Solution window not showing changes after PowerShell commands run
When installing, updating, or uninstalling a NuGet package using
one of the PowerShell commands the Solution window would not update
the packages for SDK style projects or for projects using Package
References. Now the NuGet extensions addin raises the
PackageOperationsFinished event which causes the Solution pad's
node builder to refresh the packages displayed.