Fix UI thread hang
On opening or closing a solution the NuGet PowerShell console host
calls Set-Location to set the location for the current solution or
sets it to null if the solution is closed. This code was run on the
UI thread and if the remote host does not respond this could hang
the IDE.

To fix this the PowerShell command is now run with the background
dispatcher.

Fixes #40