Improve restart logic
The restarting of the remote PowerShell host would retry again and
again until 5 seconds had passed. This was incorrect - the code
was supposed to wait for 5 seconds before retrying again.

Note that this retry/restart logic is still not great. The error
message gets displayed several times in the console output since
several commands are run on initialisation.