Fix CA1416 warning for GettextCatalog alls
CA1416 warning happens for certain VS Mac API calls or use of
Xamarin.Mac. Warning similar to the following can happen at build time
or in this case in the editor as a green wavy line:

This call site is reachable on all platforms.
'NSLayoutConstraint.Active' is only supported on: 'ios' 10.0 and later,
'maccatalyst' 10.0 and later, 'macOS/OSX' 10.14 and later, 'tvos'
10.0 and later. (CA1416))
+// Need to fix CA1416 build warning.
+// This call site is reachable on all platforms. 'NSLayoutConstraint.Ac