Fix duplicate PowerShell console prompt
When the NuGet Package Manager Console is opened after a solution
is already open then two PM> prompts are displayed. This was because
a check is made to see if any init.ps1 scripts need to be run and
when this happens to prevent any messages from this being written
to the prompt line a new line is written. This happens even if no
init.ps1 scripts are run causing an extra prompt to be displayed.

Fix this by only writing a new line if any messages are logged when
running any init.ps1 scripts. Also only show the prompt again if
any messages were logged.