Ensure dotnet is on the PATH for the remote PowerShell host
VS Mac 17.0 creates the pad windows when the IDE window is first
shown, even if the pad is never opened or available docked. This means
the PATH used to launch the powershell console host does not contain
the path to the dotnet CLI which is configured when a project is
opened. To avoid this check for a valid dotnet runtime and set the
PATH for the powershell console host to be the parent process's PATH
environment variable with the path to dotnet appended.