Fix escape codes displayed in PowerShell console
Get-Package would show ANSI escape codes for the table header. PowerShell
7 uses ANSI escape codes by default.

Setting PSStyle.Instance.OutputRendering == OutputRendering.PlainText
is not enough on its own. This only works if the last command on the
pipeline is Out-String. Using Out-Default or Out-Host as the last
command on the pipeline does not remove the ANSI escape codes.

In the PowerShellUserInterfaceHost the text is processed by the
PowerShell StringDecorated class. This removes the escape codes if
PSStyle.Instance.OutputRendering is set to be PlainText.