Re-implement parts of IVsBrowseObjectContext for .NET Core projects
If the project is an SDK style project when using EnvDTE a CpsProject
type is returned which implements the IVsBrowseObjectContext interface.
This allows EF Core tools to get a little further. EF Core's NuGet
module uses the IVsBrowseObjectContext to get the ProjectAssetsFile
path which is available as an MSBuild property.