Fix write-error output not displayed in PowerShell host
When formatting the error output PowerShell runs a script that does
this:

if ($null -ne $Host.PrivateData) {
    $errorColor = Get-VT100Color $Host.PrivateData.ErrorForegroundColor
    $accentColor = Get-VT100Color ($Host.PrivateData.ErrorAccentColor ?? $errorColor)
}

Since the PrivateData is non-null this code was failing since
the PrivateData object did not implement the various ConsoleColor
properties that PowerShell assumes are available there. Fix this
by adding the various ConsoleColor properties which currently do
nothing.